pyproject.toml
[tool.poetry]
name = "ssef-bot"
version = "0.1.0"
description = "SSEF's Bot"
authors = ["Vioshim"]
readme = "README.md"
packages = [{include = "ssef_bot"}]

[tool.poetry.dependencies]
python = "^3.11.4"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", rev = "master", extras = [
    "speed",
    "voice",
] }
jishaku = "^2.5.1"
uvloop = { version = "^0.17.0", platform = "linux" }
rapidfuzz = "^2.6.0"
motor = { extras = ["srv"], version = "^3.0.0" }
pymongo = { extras = ["srv"], version = "^4.2.0" }
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^6.0.0"
motor-stubs = "^1.7.1"
black = { version = "^23.1.0", allow-prereleases = true }
isort = { version = "^5.10.1", extras = [
    "requirements_deprecated_finder",
    "pipfile_deprecated_finder",
] }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]

line-length = 120 # override black's default line-length
target-version = ["py311"]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | .vscode
  | .history
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.pyright]
useLibraryCodeForTypes = true
typeCheckingMethod = "basic"
pythonVersion = "3.11.4"
flake8 = "^4.0.1"


[tool.isort]
# make it compatible with black
profile = "black"
